% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g_mini}
\alias{g_mini}
\title{Example data for the rsemmed package}
\format{
An \code{igraph} with 7 nodes and 15 edges
}
\usage{
data(g_mini)
}
\description{
A dataset containing a very small subset of the full 
Semantic MEDLINE graph.
}
\keyword{datasets}
