% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g_small}
\alias{g_small}
\title{Example data for the rsemmed package}
\format{
An \code{igraph} with 1038 nodes and 318,105 edges
}
\usage{
data(g_small)
}
\description{
A dataset containing a small subset of the full 
Semantic MEDLINE graph.
}
\keyword{datasets}
